+++ example/lib/main.dart
@@ -1,3 +1,9 @@
+/*
+*
+* Check live demo with various use-cases and its code on https://fluttercommunity.github.io/backdrop/demo/#/
+*
+* */
+
 import 'package:backdrop/backdrop.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_platform_widgets/flutter_platform_widgets.dart';