+++ lib/scaffold.dart
@@ -63,8 +63,14 @@ class Backdrop extends InheritedWidget {
 /// See also:
 ///  * [Scaffold], which is the plain scaffold used in material apps.
 class BackdropScaffold extends StatefulWidget {
+  /// Deprecated. Use [animationController].
+  ///
   /// Can be used to customize the behaviour of the backdrop animation.
+  @Deprecated("See animationController. This was deprecated after v0.5.1")
   final AnimationController controller;
+
+  /// Can be used to customize the behaviour of the backdrop animation.
+  final AnimationController animationController;
 
   /// Deprecated. Use [BackdropAppBar.title].
   ///
@@ -77,10 +83,16 @@ class BackdropScaffold extends StatefulWidget {
   /// The widget that is shown on the front layer.
   final Widget frontLayer;
 
-  /// The widget that is shown as sub-header on top of the front layer.
+  /// The widget shown at the top of the front layer.
+  ///
+  /// When the front layer is minimized (back layer revealed), the entire [subHeader] will be visible unless
+  /// [headerHeight] is specified.
   final Widget subHeader;
 
-  /// This boolean flag keeps subHeader active when [backLayer] is visible. Defaults to true.
+  /// If true, the scrim applied to the front layer while minimized (back layer revealed) will not
+  /// cover the [subHeader].  See [frontLayerScrim].
+  ///
+  /// Defaults to true.
   final bool subHeaderAlwaysActive;
 
   /// Deprecated. Use [BackdropAppBar.actions].
@@ -88,19 +100,15 @@ class BackdropScaffold extends StatefulWidget {
   /// Actions passed to [AppBar.actions].
   final List<Widget> actions;
 
-  /// Defines the height of the front layer when it is in the opened state.
+  /// Defines the front layer's height when minimized (back layer revealed)).
+  /// Defaults to measured height of [subHeader] if provided, else 32.
   ///
-  /// This height value is only applied, if [stickyFrontLayer]
-  /// is set to `false` or if [stickyFrontLayer] is set to `true` and the back
-  /// layer's height is less than
-  /// [BoxConstraints.biggest.height]-[headerHeight].
+  /// To automatically use the difference of the screen height and back layer's height,
+  /// see [stickyFrontLayer].  Note [headerHeight] is ignored if it is less
+  /// than the available size and [stickyFrontLayer] is `true`.
   ///
-  /// [headerHeight] is interpreted as the height of the front layer's visible
-  /// part, when being opened. The back layer's height corresponds to
-  /// [BoxConstraints.biggest.height]-[headerHeight].
-  ///
-  ///
-  /// If [subHeader] is defined then height of subHeader otherwise defaults to 32.0.
+  /// To vary the front layer's height when active (back layer concealed),
+  /// see [frontLayerActiveFactor].
   final double headerHeight;
 
   /// Defines the [BorderRadius] applied to the front layer.
@@ -108,8 +116,8 @@ class BackdropScaffold extends StatefulWidget {
   /// Defaults to
   /// ```dart
   /// const BorderRadius.only(
-  ///     topLeft: Radius.circular(16.0),
-  ///     topRight: Radius.circular(16.0),
+  ///     topLeft: Radius.circular(16),
+  ///     topRight: Radius.circular(16),
   /// )
   /// ```
   final BorderRadius frontLayerBorderRadius;
@@ -117,14 +125,28 @@ class BackdropScaffold extends StatefulWidget {
   /// A flag indicating whether the front layer should stick to the height of
   /// the back layer when being opened.
   ///
-  /// Defaults to `false`.
+  /// This parameter has no effect if the back layer's measured height
+  /// is greater than or equal to the screen height.
   final bool stickyFrontLayer;
 
-  /// The animation curve passed to [Tween.animate]() when triggering
+  /// Flag indicating whether the back layer should be revealed at the beginning
+  /// or not. Setting [revealBackLayerAtStart] to `true` reveals the back layer
+  /// at start. This property has no effect if a custom [animationController]
+  /// is set.
+  ///
+  /// Defaults to `false`.
+  final bool revealBackLayerAtStart;
+
+  /// The animation curve passed to [Tween.animate] when triggering
   /// the backdrop animation.
   ///
-  /// Defaults to [Curves.easeInOut].
+  /// Defaults to [Curves.ease].
   final Curve animationCurve;
+
+  /// The reverse animation curve passed to [Tween.animate].
+  ///
+  /// If not set, [animationCurve.flipped] is used.
+  final Curve reverseAnimationCurve;
 
   /// Background [Color] for the back layer.
   ///
@@ -136,27 +158,41 @@ class BackdropScaffold extends StatefulWidget {
   /// If null, the color is handled automatically according to the theme.
   final Color frontLayerBackgroundColor;
 
-  /// Defines the color for the inactive front layer.
-  /// A default opacity of 0.7 is applied to the passed color.
-  /// See [inactiveOverlayOpacity].
+  /// Fraction of the available height the front layer will occupy,
+  /// when active (back layer concealed).  Clamped to (0, 1).
   ///
-  /// Defaults to `const Color(0xFFEEEEEE)`.
+  /// Note the front layer will not fully conceal the back layer when
+  /// this value is less than 1.  A scrim will cover the
+  /// partially concealed back layer; see [backLayerScrim].
+  ///
+  /// Defaults to 1.
+  final double frontLayerActiveFactor;
+
+  /// Deprecated.  Use [frontLayerScrim] instead.
+  @Deprecated('Replace with frontLayerScrim.')
   final Color inactiveOverlayColor;
 
-  /// The opacity value applied to [inactiveOverlayColor] when used on the
-  /// inactive layer.
-  ///
-  /// The inactive layer overlays the [frontLayer] when the back layer is
-  /// revealed and the front layer is in a concealed state.
-  ///
-  /// Must be a value between `0.0` and `1.0`.
-  /// Defaults to `0.7`.
+  /// Deprecated.  Use [frontLayerScrim] instead.
+  @Deprecated('Replace with frontLayerScrim.  Use Color#withOpacity, or pass'
+      'the opacity value in the Color constructor.')
   final double inactiveOverlayOpacity;
 
-  /// Will be called when [backLayer] have been concealed.
+  /// Defines the scrim color for the front layer when minimized
+  /// (revealing the back layer) and animating.  Defaults to [Colors.white70].
+  ///
+  /// See [subHeaderAlwaysActive] to leave the [subHeader] outside the scrim.
+  final Color frontLayerScrim;
+
+  /// Defines the scrim color for the back layer when partially concealed
+  /// (with [frontLayerActiveFactor] less than 1).
+  ///
+  /// Defaults to [Colors.black54].
+  final Color backLayerScrim;
+
+  /// Will be called when [backLayer] has been concealed.
   final VoidCallback onBackLayerConcealed;
 
-  /// Will be called when [backLayer] have been revealed.
+  /// Will be called when [backLayer] has been revealed.
   final VoidCallback onBackLayerRevealed;
 
   // ------------- PROPERTIES TAKEN OVER FROM SCAFFOLD ------------- //
@@ -178,7 +214,9 @@ class BackdropScaffold extends StatefulWidget {
   /// Creates a backdrop scaffold to be used as a material widget.
   BackdropScaffold({
     Key key,
-    this.controller,
+    @Deprecated("See animationController. This was deprecated after v0.5.1")
+        this.controller,
+    this.animationController,
     @Deprecated("Replace by use of BackdropAppBar. See BackdropAppBar.title."
         "This feature was deprecated after v0.2.17.")
         this.title,
@@ -193,20 +231,28 @@ class BackdropScaffold extends StatefulWidget {
         this.actions = const <Widget>[],
     this.headerHeight,
     this.frontLayerBorderRadius = const BorderRadius.only(
-      topLeft: Radius.circular(16.0),
-      topRight: Radius.circular(16.0),
+      topLeft: Radius.circular(16),
+      topRight: Radius.circular(16),
     ),
     this.stickyFrontLayer = false,
-    this.animationCurve = Curves.easeInOut,
+    this.revealBackLayerAtStart = false,
+    this.animationCurve = Curves.ease,
+    this.reverseAnimationCurve,
     this.frontLayerBackgroundColor,
+    double frontLayerActiveFactor = 1,
     this.backLayerBackgroundColor,
-    this.inactiveOverlayColor = const Color(0xFFEEEEEE),
-    this.inactiveOverlayOpacity = 0.7,
+    @Deprecated('See frontLayerScrim. This was deprecated after v0.4.7.')
+        this.inactiveOverlayColor = const Color(0xFFEEEEEE),
+    @Deprecated('See frontLayerScrim. This was deprecated after v0.4.7.')
+        this.inactiveOverlayOpacity = 0.7,
+    this.frontLayerScrim = Colors.white70,
+    this.backLayerScrim = Colors.black54,
     this.onBackLayerConcealed,
     this.onBackLayerRevealed,
     this.materialScaffoldData,
     this.cupertinoPageScaffoldData,
   })  : assert(inactiveOverlayOpacity >= 0.0 && inactiveOverlayOpacity <= 1.0),
+        frontLayerActiveFactor = frontLayerActiveFactor.clamp(0, 1).toDouble(),
         super(key: key);
 
   @override
@@ -223,8 +269,9 @@ class BackdropScaffold extends StatefulWidget {
 /// within the widget tree below [BackdropScaffold].
 class BackdropScaffoldState extends State<BackdropScaffold>
     with SingleTickerProviderStateMixin {
-  bool _shouldDisposeController = false;
-  AnimationController _controller;
+  bool _shouldDisposeAnimationController = false;
+  AnimationController _animationController;
+  ColorTween _backLayerScrimColorTween;
 
   /// Key for accessing the [ScaffoldState] of [BackdropScaffold]'s internally
   /// used [Scaffold].
@@ -232,14 +279,21 @@ class BackdropScaffoldState extends State<BackdropScaffold>
   double _backPanelHeight = 0;
   double _subHeaderHeight = 0;
 
+  /// Deprecated. Use [animationController] instead.
+  ///
+  /// [AnimationController] used for the backdrop animation.
+  @Deprecated("Replace by the use of `animationController`."
+      "This feature was deprecated after v0.5.1.")
+  AnimationController get controller => _animationController;
+
   /// [AnimationController] used for the backdrop animation.
   ///
   /// Defaults to
   /// ```dart
   /// AnimationController(
-  ///         vsync: this, duration: Duration(milliseconds: 200), value: 1.0)
+  ///         vsync: this, duration: Duration(milliseconds: 200), value: 1)
   /// ```
-  AnimationController get controller => _controller;
+  AnimationController get animationController => _animationController;
 
   @override
   void initState() {
@@ -247,39 +301,53 @@ class BackdropScaffoldState extends State<BackdropScaffold>
     // initialize scaffoldKey
     scaffoldKey = widget.scaffoldKey ?? GlobalKey<ScaffoldState>();
     // initialize _controller
-    _controller = widget.controller ??
+    _animationController = widget.animationController ??
+        widget.controller ??
         AnimationController(
-            vsync: this, duration: Duration(milliseconds: 200), value: 1.0);
-    if (widget.controller == null) _shouldDisposeController = true;
+          vsync: this,
+          duration: Duration(milliseconds: 200),
+          value: widget.revealBackLayerAtStart ? 0 : 1,
+        );
+    if (widget.animationController == null && widget.controller == null) {
+      _shouldDisposeAnimationController = true;
+    }
 
-    _controller.addStatusListener((status) {
-      setState(() {
-        // This is intentionally left empty. The state change itself takes
-        // place inside the AnimationController, so there's nothing to update.
-        // All we want is for the widget to rebuild and read the new animation
-        // state from the AnimationController.
-        // see https://github.com/flutter/flutter/pull/55414/commits/72d7d365be6639271a5e88ee3043b92833facb79
-      });
-    });
+    _backLayerScrimColorTween = _buildBackLayerScrimColorTween();
+
+    _animationController.addListener(() => setState(() {
+          // This is intentionally left empty. The state change itself takes
+          // place inside the AnimationController, so there's nothing to update.
+          // All we want is for the widget to rebuild and read the new animation
+          // state from the AnimationController.
+          // see https://github.com/flutter/flutter/pull/55414/commits/72d7d365be6639271a5e88ee3043b92833facb79
+        }));
+  }
+
+  @override
+  void didUpdateWidget(covariant BackdropScaffold oldWidget) {
+    super.didUpdateWidget(oldWidget);
+    if (oldWidget.backLayerScrim != widget.backLayerScrim) {
+      _backLayerScrimColorTween = _buildBackLayerScrimColorTween();
+    }
   }
 
   @override
   void dispose() {
     super.dispose();
-    if (_shouldDisposeController) _controller.dispose();
+    if (_shouldDisposeAnimationController) _animationController.dispose();
   }
 
   /// Deprecated. Use [isBackLayerConcealed] instead.
   ///
-  /// Wether the back layer is concealed or not.
+  /// Whether the back layer is concealed or not.
   @Deprecated("Replace by the use of `isBackLayerConcealed`."
       "This feature was deprecated after v0.3.2.")
   bool get isTopPanelVisible => isBackLayerConcealed;
 
-  /// Wether the back layer is concealed or not.
+  /// Whether the back layer is concealed or not.
   bool get isBackLayerConcealed =>
-      controller.status == AnimationStatus.completed ||
-      controller.status == AnimationStatus.forward;
+      animationController.status == AnimationStatus.completed ||
+      animationController.status == AnimationStatus.forward;
 
   /// Deprecated. Use [isBackLayerRevealed] instead.
   ///
@@ -290,8 +358,8 @@ class BackdropScaffoldState extends State<BackdropScaffold>
 
   /// Whether the back layer is revealed or not.
   bool get isBackLayerRevealed =>
-      controller.status == AnimationStatus.dismissed ||
-      controller.status == AnimationStatus.reverse;
+      animationController.status == AnimationStatus.dismissed ||
+      animationController.status == AnimationStatus.reverse;
 
   /// Toggles the backdrop functionality.
   ///
@@ -317,7 +385,7 @@ class BackdropScaffoldState extends State<BackdropScaffold>
   /// Animates the back layer to the "revealed" state.
   void revealBackLayer() {
     if (isBackLayerConcealed) {
-      controller.animateBack(-1.0);
+      animationController.animateBack(-1);
       widget.onBackLayerRevealed?.call();
     }
   }
@@ -332,7 +400,7 @@ class BackdropScaffoldState extends State<BackdropScaffold>
   /// Animates the back layer to the "concealed" state.
   void concealBackLayer() {
     if (isBackLayerRevealed) {
-      controller.animateTo(1.0);
+      animationController.animateTo(1);
       widget.onBackLayerConcealed?.call();
     }
   }
@@ -341,8 +409,8 @@ class BackdropScaffoldState extends State<BackdropScaffold>
     // if defined then use it
     if (widget.headerHeight != null) return widget.headerHeight;
 
-    // if no subHeader then 32.0
-    if (widget.subHeader == null) return 32.0;
+    // if no subHeader then 32
+    if (widget.subHeader == null) return 32;
 
     // if subHeader then height of subHeader
     return _subHeaderHeight;
@@ -351,46 +419,46 @@ class BackdropScaffoldState extends State<BackdropScaffold>
   Animation<RelativeRect> _getPanelAnimation(
       BuildContext context, BoxConstraints constraints) {
     double backPanelHeight, frontPanelHeight;
-
-    if (widget.stickyFrontLayer &&
-        _backPanelHeight < constraints.biggest.height - _headerHeight) {
+    final availableHeight = constraints.biggest.height - _headerHeight;
+    if (widget.stickyFrontLayer && _backPanelHeight < availableHeight) {
       // height is adapted to the height of the back panel
       backPanelHeight = _backPanelHeight;
       frontPanelHeight = -_backPanelHeight;
     } else {
       // height is set to fixed value defined in widget.headerHeight
-      final height = constraints.biggest.height;
-      backPanelHeight = height - _headerHeight;
+      backPanelHeight = availableHeight;
       frontPanelHeight = -backPanelHeight;
     }
     return RelativeRectTween(
-      begin: RelativeRect.fromLTRB(0.0, backPanelHeight, 0.0, frontPanelHeight),
-      end: RelativeRect.fromLTRB(0.0, 0.0, 0.0, 0.0),
+      begin: RelativeRect.fromLTRB(0, backPanelHeight, 0, frontPanelHeight),
+      end: RelativeRect.fromLTRB(
+          0, availableHeight * (1 - widget.frontLayerActiveFactor), 0, 0),
     ).animate(CurvedAnimation(
-      parent: controller,
-      curve: widget.animationCurve,
-    ));
+        parent: animationController,
+        curve: widget.animationCurve,
+        reverseCurve:
+            widget.reverseAnimationCurve ?? widget.animationCurve.flipped));
   }
 
   Widget _buildInactiveLayer(BuildContext context) {
     return Offstage(
-      offstage: controller.status == AnimationStatus.completed,
+      offstage: animationController.status == AnimationStatus.completed,
       child: FadeTransition(
-        opacity: Tween(begin: 1.0, end: 0.0).animate(controller),
+        opacity: Tween<double>(begin: 1, end: 0).animate(animationController),
         child: GestureDetector(
           onTap: () => fling(),
           behavior: HitTestBehavior.opaque,
           child: Column(
             children: <Widget>[
-              // if subHeaderAlwaysActive then do not apply inactiveOverlayColor for area with _headerHeight
+              // if subHeaderAlwaysActive then do not apply frontLayerScrim for area with _subHeaderHeight
               widget.subHeader != null && widget.subHeaderAlwaysActive
-                  ? Container(height: _headerHeight)
+                  ? Container(height: _subHeaderHeight)
                   : Container(),
               Expanded(
                 child: Container(
-                  color: widget.inactiveOverlayColor
-                      .withOpacity(widget.inactiveOverlayOpacity),
-                ),
+                    color: widget.frontLayerScrim ??
+                        widget.inactiveOverlayColor
+                            .withOpacity(widget.inactiveOverlayOpacity)),
               ),
             ],
           ),
@@ -400,30 +468,35 @@ class BackdropScaffoldState extends State<BackdropScaffold>
   }
 
   Widget _buildBackPanel() {
-    return FocusScope(
-      canRequestFocus: isBackLayerRevealed,
-      child: Material(
-        color: this.widget.backLayerBackgroundColor ??
-            Theme.of(context).primaryColor,
-        child: Column(
-          children: <Widget>[
-            Flexible(
-              child: _MeasureSize(
-                onChange: (size) =>
-                    setState(() => _backPanelHeight = size.height),
-                child: widget.backLayer ?? Container(),
-              ),
+    return Stack(
+      children: [
+        FocusScope(
+          canRequestFocus: isBackLayerRevealed,
+          child: Material(
+            color: widget.backLayerBackgroundColor ??
+                Theme.of(context).primaryColor,
+            child: Column(
+              children: <Widget>[
+                Flexible(
+                  child: _MeasureSize(
+                    onChange: (size) =>
+                        setState(() => _backPanelHeight = size.height),
+                    child: widget.backLayer ?? Container(),
+                  ),
+                ),
+              ],
             ),
-          ],
+          ),
         ),
-      ),
+        if (_hasBackLayerScrim) _buildBackLayerScrim(),
+      ],
     );
   }
 
   Widget _buildFrontPanel(BuildContext context) {
     return Material(
-      color: this.widget.frontLayerBackgroundColor,
-      elevation: 1.0,
+      color: widget.frontLayerBackgroundColor,
+      elevation: 1,
       borderRadius: widget.frontLayerBorderRadius,
       child: ClipRRect(
         borderRadius: widget.frontLayerBorderRadius,
@@ -461,6 +534,11 @@ class BackdropScaffoldState extends State<BackdropScaffold>
     return true;
   }
 
+  ColorTween _buildBackLayerScrimColorTween() => ColorTween(
+        begin: Colors.transparent,
+        end: widget.backLayerScrim,
+      );
+
   Widget _buildBody(BuildContext context) {
     return WillPopScope(
       onWillPop: () => _willPopCallback(context),
@@ -490,6 +568,13 @@ class BackdropScaffoldState extends State<BackdropScaffold>
     );
   }
 
+  Container _buildBackLayerScrim() => Container(
+      color: _backLayerScrimColorTween.evaluate(animationController),
+      height: _backPanelHeight);
+
+  bool get _hasBackLayerScrim =>
+      isBackLayerConcealed && widget.frontLayerActiveFactor < 1;
+
   @override
   Widget build(BuildContext context) {
     return Backdrop(
@@ -500,7 +585,6 @@ class BackdropScaffoldState extends State<BackdropScaffold>
     );
   }
 }
-
 
 /// Widget to get size of child widget
 /// Credit: https://stackoverflow.com/a/60868972/2554745